gave a cheat radius for being hit
circles of enemy or bullet has to overlap the circle of the player
by a certain amount in order for a hit to be registered. (this is a fix
of a regression that happened with the powerup merge.) no such check
is performed against the enemies, so they are still as easy to hit.
fixed problem of body of shooter not doing damage