smarter hit delay on beams
Beams now store the enemies they've hit in a map, and tick down a
cooldown for each of them every step. This means that a creature always
gets hit the first time it enters a beam, and it rate limits beam damage
in a more sane way. It also scales with bulletSpeed so that attribute
isn't completely useless once you get beam